separate coponents
